+++ spec/services/stellar_base/account_subscriptions/subscribe_account_spec.rb
@@ -14,7 +14,11 @@ module StellarBase
       expect(StellarBase.configuration.on_account_event).to receive(:call).
         exactly(10).times.
         and_call_original
-      described_class.(account_subscription, operation_limit: 10)
+      result = described_class.(account_subscription, operation_limit: 10)
+
+      expect(result).to be_success
+      expect(result.operations.first).to be_a StellarOperation
+      expect(result.tx).to be_a StellarTransaction
     end
 
   end